Refactor Android implementation to not use singleton
The Segment documentation recommends using the singleton, but for React Native apps,
this does not really work well. Reloading the app, for example, could trigger the setup method a second time,
which would crash the app.
So instead, we will keep a reference to the analytics object in the module, and update it as necessary.