Refactor Android implementation to use singleton (again)
The Segment SDK gets very unhappy when we instantiate a second Analytics object,
so instead, we go back to using the provided singleton. The `setup` method can be
called once to perform the initial setup, and subsequent calls will result in failure.
That will, however, be wrapped in a promise, so it does not crash the app that is
using the module. This means that it is now up to the consumer of the module to
deal with success and failure.