Add debug option to setup to log during development
It's nice to log to the console/logcat while developing, to verify that
everything works as expected. You can just pass `debug` as one of the
elements in the options map, and (if set to `true`) it will set the
log level to `VERBOSE` on Android and the `debug` flag to `YES` on iOS.

This is not logging in the JavaScript console, but the iOS dev console
and Android's logcat (just to make this perfectly clear).